src/favorites/favorites.controller.ts
import { Controller, Get, Post, Body, Patch, Param, Delete, Logger, Query } from '@nestjs/common';
import { FavoritesService } from './favorites.service';
import { CreateFavoriteDto } from './dto/create-favorite.dto';

@Controller('favorites')
export class FavoritesController {
  constructor(private readonly favoritesService: FavoritesService) { }

  @Post()
  create(@Body() createFavoriteDto: CreateFavoriteDto) {
    return this.favoritesService.create(createFavoriteDto);
  }


  @Delete()
  remove(
    @Query('user_id') userId: string,
    @Query('cat_id') catId: string
  ) {
    return this.favoritesService.remove(userId, catId);
  }
}